Classes and Objects OOP_PRPROB15
An office management system was designed through a java program for storing the records of employees. The program should accept the following details from the user: Employee id, name, years of employment, total salaries in the past years of employment. Define the methods complete_sal() and overall_sal() which should calculate the total salaries and average salaries of the employee.

Input Format
1011 Hitesh 3 11000 54000 3000

Output Format
78000 26000