DS_QUEUE01 Implement Queue using Array
A queue of people is waiting for a ticket at BRTS Ticket counter. Any new person is joining at one end of the queue, you can call it the rear end. When the chance arrives the person at the other end first buys the ticket, you can call it as the front end of the queue. A person who got the ticket left the queue. There is the maximum capacity of the queue is 5.

Note:- If the queue is empty and you are removing the person from the queue then it should display "Queue is Empty", if it is reached the capacity(at rear end) and if you want to add more people in the queue then it should display "Queue is Full"

Three persons join in the queue
Two persons removed from the queue
One person joins in the queue
One person removed from the queue
One person joins in the queue
One person joins in the queue
Input Format
10 20 30 40 50 60

Output Format
40 50
Queue is Full