age is more than or equal to 18
Person has attributes name, age and a method to verify his age for voting which will throw an exception if he/she is under-aged (i.e. less than 18) for voting. In such case, application class must handle the exception. If age is more than or equal to 18, the program will run smoothly.