calculate the total amount to be paid
Mayra is developing an application related to the banking sector, where she is required to take input like ROI which is in float format, principal which is in INT format, no_of_years in float format. Make sure the exceptions handle individual formats. If input is correct, calculate the total amount to be paid using compound interest.

NOTE: if no_of_years is 2.12 convert it to 3 years(handle in the exception)