Display overall list of elements present within queue.
Rajiv was assigned a task to structure the data using a Queue interface wherein he was asked to structure the programming subjects semester wise within the queue followed by retrieval of the head element of the queue. Remove the two heads of the queue and add an element as ‘Semester 6: Pascal’ into the queue.
Display the overall list of elements present within the queue.