Dollar, Cents : (Split funtion)
1. You are developing a program that classifies a given amount of money into
smaller monetary units. The program lets the user enter an amount
representing a total in dollars and cents, and then outputs a report listing the
monetary equivalent in dollars, quarters, dimes, nickels, and pennies, as
shown in the sample run. Your program should report the maximum number
of dollars, then the number of quarters, dimes, nickels, and pennies, in this
order, to result in the minimum number of coins.
Here is a sample run:
Enter an amount in double, for example 11.56: 11.56
Your amount 11.56 consists of
11 dollars
2 quarters
0 dimes
1 nickels
1 pennies