Random Number: Lottery 10k, 5k, 2k, 1k.
2. Suppose you want to develop a program to play a lottery. The program
randomly generates a two-digit number, prompts the user to enter a two-
digit number, and determines whether the user wins according to the
following rules:
a. If the user’s input matches the lottery in the exact order, the award is
$10,000.
b. If all the digits in the user’s input match all the digits in the lottery
number, the award is $5,000.
c. If one digit in the user’s input matches a digit in the lottery number,
the award is $2,000